+++ Assets/Xsolla/Core/Editor/AndroidManifestUtils/FindCriterias/FindByChildName.cs
@@ -1,0 +1,32 @@
+﻿using System.Xml;
+
+namespace Xsolla.Core
+{
+	public class FindByChildName : IFindCriteria<XmlNode>
+	{
+		private readonly string tag;
+		private readonly BaseManifestNode childNode;
+
+		public FindByChildName(string tag, BaseManifestNode childNode)
+		{
+			this.tag = tag;
+			this.childNode = childNode;
+		}
+
+		public bool MatchesCriteria(XmlNode xmlNode)
+		{
+			if (xmlNode == null)
+			{
+				return false;
+			}
+
+			if (tag.Equals(xmlNode.Name))
+			{
+				var child = xmlNode.FindNodeRecursive(new FindByName(this.childNode));
+				return child != null;
+			}
+
+			return false;
+		}
+	}
+}