+++ Assets/Xsolla/Core/Editor/AndroidManifestUtils/ManifestNodes/ActivityNode.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace Xsolla.Core
+{
+	public class ActivityNode : BaseManifestNode
+	{
+		public ActivityNode(string name) : base(
+			AndroidManifestConstants.ActivityTag, AndroidManifestConstants.ApplicationTag,
+			new Dictionary<string, string> {{AndroidManifestConstants.NameAttribute, name}})
+		{
+		}
+	}
+}