+++ Assets/Xsolla/Scripts/Core/XsollaSettings/Editor/XsollaSettingsEditor.cs
@@ -1,4 +1,4 @@
-﻿using UnityEditor;
+using UnityEditor;
 using Xsolla.Login;
 
 namespace Xsolla.Core
@@ -14,13 +14,13 @@ namespace Xsolla.Core
 
 		public override void OnInspectorGUI()
 		{
-			var changed = XsollaLoginSettings();
+			var changed =   XsollaLoginSettings() ||
+                            PublishingPlatformSettings() ||
+			                XsollaStoreSettings() ||
+                            XsollaPaystationSettings() ||
+                            AndroidSDKSettings();
 
-			changed = changed || PublishingPlatformSettings();
-
-			changed = changed || XsollaStoreSettings();
-
-			XsollaPaystationSettings();
+			XsollaAndroidSettings();
 
 			if (changed)
 			{